refactor: StorageService 추상화
