+++ backend/src/test/java/com/zzang/chongdae/offering/util/FakeStorageService.java
@@ -1,0 +1,12 @@
+package com.zzang.chongdae.offering.util;
+
+import com.zzang.chongdae.storage.service.StorageService;
+import org.springframework.web.multipart.MultipartFile;
+
+public class FakeStorageService implements StorageService {
+
+    @Override
+    public String uploadFile(MultipartFile file, String path) {
+        return "https://upload-image-url.com/";
+    }
+}
